feat(cli): complete tui-sidebar-live — live check, verify fixes, specs synced
Closes out change 2 of doc 14 (the checkpoint commit landed the implementation):

- live-verified via tmux screen capture (PTY stream matching is unreliable for
  opentui cell-diff paints — recorded as a method finding): DATA PROFILE renders
  the real completed profile from the ledger, RUNS its truthful empty state, and
  ctrl+x d opens the details dialog with the agent's real summary
- the live pass caught a real integration bug, fixed with tests: the parity
  trigger is the one lifecycle edge that mutates ledger state outside the
  sidebar's refresh triggers, so its triggered outcome now pokes the store
  (trigger -> refresh -> arm -> flip chain unit-tested link by link)
- adversarial verify (SHIP, 0 critical) warnings all fixed: honest RunBlock
  footer via a hint prop, run-row helpers single-homed, profileDetailLines moved
  to sidebar_live with the gallery calling the real function over a fixture,
  unavailable snapshots keep the poll alive (transient DB blips self-heal),
  comment/spec-example hygiene, warn tone for suspended_insufficient_funds
- three delta specs synced (sidebar-live created; tui-layout sidebar requirement
  goes live-data; tui-mock-data becomes gallery-only) and the change archived

573 cli tests pass (0 fail); typecheck + eslint clean.
