+++ cli/openspec/changes/archive/2026-07-08-tui-sidebar-live/specs/tui-mock-data/spec.md
