+++ cli/src/tui/hooks/sidebar_live.test.ts
@@ -2,6 +2,9 @@ import { afterEach, describe, expect, test } from "bun:test";
 import { ok, okAsync, errAsync, ResultAsync } from "neverthrow";
 import { createRoot } from "solid-js";
 
+// Side-effect import: installs `Date.relativeAge` (the loaded-profile timestamp lines call it) via the
+// same central loader the app boots with.
+import "../../extensions/index.ts";
 import type { CortexRunRow, DataProfileStatus, DbError } from "@inflexa-ai/harness";
 import type { ResolvedHarnessConfig } from "../../modules/harness/config.ts";
 import type { HarnessRuntime } from "../../modules/harness/runtime.ts";
@@ -11,6 +14,7 @@ import { setChatStatus } from "./status.ts";
 import {
     __resetSidebarLiveForTest,
     hasActiveWork,
+    profileDetailLines,
     profileSnapshot,
     refreshSidebarData,
     runsSnapshot,
@@ -83,6 +87,30 @@ function mountWatch(ws: Workspace, watchSeams: WatchSeams): () => void {
 function wsFor(id: string | null): Workspace {
     const analysis = id === null ? null : ({ id } as unknown as Workspace["analysis"]);
     return { analysis } as unknown as Workspace;
+}
+
+/** Build a `loaded` profile snapshot for the {@link profileDetailLines} composer tests. */
+function loaded(over: Partial<DataProfileStatus> = {}): ProfileSnapshot {
+    return {
+        kind: "loaded",
+        profile: {
+            status: "completed",
+            error: null,
+            startedAt: "2026-07-08T00:00:00.000Z",
+            completedAt: "2026-07-08T00:00:05.000Z",
+            result: {
+                summary: "line one\nline two",
+                files: [
+                    { path: "data/counts.tsv", description: "raw counts" },
+                    { path: "data/meta.csv", description: "sample metadata" },
+                ],
+                inputFileIds: ["i1", "i2"],
+                profiledAt: "2026-07-08T00:00:05.000Z",
+            },
+            seedInputFileIds: ["i1", "i2", "i3"],
+            ...over,
+        },
+    };
 }
 
 describe("refreshSidebarData — snapshot ladder", () => {
@@ -186,9 +214,14 @@ describe("hasActiveWork — poll arming predicate", () => {
         expect(hasActiveWork({ kind: "absent" }, { kind: "loaded", runs: terminal })).toBe(false);
     });
 
-    test("not_ready / unavailable snapshots are never active (idle costs nothing)", () => {
+    test("not_ready snapshots alone are never active (idle costs nothing)", () => {
         expect(hasActiveWork(notReady, { kind: "not_ready" })).toBe(false);
-        expect(hasActiveWork({ kind: "unavailable" }, { kind: "unavailable" })).toBe(false);
+    });
+
+    test("an unavailable snapshot arms — a transient DB blip self-heals via the same 5s poll", () => {
+        expect(hasActiveWork({ kind: "unavailable" }, { kind: "not_ready" })).toBe(true);
+        expect(hasActiveWork(notReady, { kind: "unavailable" })).toBe(true);
+        expect(hasActiveWork({ kind: "unavailable" }, { kind: "unavailable" })).toBe(true);
     });
 });
 
@@ -271,5 +304,54 @@ describe("watchSidebarData — triggers and bounded poll", () => {
         expect(disarms).toBe(0);
         dispose();
         expect(disarms).toBe(1); // onCleanup disarmed the live interval
+    });
+});
+
+describe("profileDetailLines — one line set per snapshot kind", () => {
+    test("not_ready → a single placeholder line", () => {
+        expect(profileDetailLines({ kind: "not_ready" })).toEqual(["runtime not ready"]);
+    });
+
+    test("absent → not profiled yet", () => {
+        expect(profileDetailLines({ kind: "absent" })).toEqual(["not profiled yet"]);
+    });
+
+    test("unavailable → status unavailable", () => {
+        expect(profileDetailLines({ kind: "unavailable" })).toEqual(["profile status unavailable"]);
+    });
+
+    test("loaded completed → status, times, summary, per-file, seed count", () => {
+        const lines = profileDetailLines(loaded());
+        expect(lines[0]).toBe("status: completed");
+        expect(lines.some((l) => l.startsWith("started "))).toBe(true);
+        expect(lines.some((l) => l.startsWith("completed "))).toBe(true);
+        expect(lines).toContain("line one");
+        expect(lines).toContain("line two");
+        expect(lines).toContain("files (2):");
+        expect(lines.some((l) => l.includes("data/counts.tsv") && l.includes("raw counts"))).toBe(true);
+        expect(lines.some((l) => l.includes("data/meta.csv") && l.includes("sample metadata"))).toBe(true);
+        // seedInputFileIds (3) wins over the profiled inputFileIds count.
+        expect(lines[lines.length - 1]).toBe("3 seed inputs");
+    });
+
+    test("loaded failed → surfaces the multi-line error", () => {
+        const lines = profileDetailLines(loaded({ status: "failed", error: "boom\ndetails here", result: null, seedInputFileIds: null }));
+        expect(lines[0]).toBe("status: failed");
+        expect(lines).toContain("boom");
+        expect(lines).toContain("details here");
+        // No result + no seed set → zero, pluralized.
+        expect(lines[lines.length - 1]).toBe("0 seed inputs");
+    });
+
+    test("loaded pending without a result → status + seed count, no files section", () => {
+        const lines = profileDetailLines(
+            loaded({ status: "pending", startedAt: "2026-07-08T00:00:00.000Z", completedAt: null, result: null, seedInputFileIds: ["only-one"] }),
+        );
+        expect(lines[0]).toBe("status: pending");
+        expect(lines.some((l) => l.startsWith("started "))).toBe(true);
+        expect(lines.some((l) => l.startsWith("completed "))).toBe(false);
+        expect(lines.some((l) => l.startsWith("files ("))).toBe(false);
+        // Singular when exactly one seed input.
+        expect(lines[lines.length - 1]).toBe("1 seed input");
     });
 });
